+++ src/Dto/In/CompanyDtoIn.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Dto;
+namespace App\Dto\In;
 
 class CompanyDtoIn
 {
