+++ src/Dto/In/InvoiceDto.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Dto;
+namespace App\Dto\In;
 
 use DateTimeInterface;
