+++ src/Dto/In/TariffDtoIn.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Dto;
+namespace App\Dto\In;
 
 class TariffDtoIn
 {
